IoTServicesCode/message_router/requirements.txt
Flask
Flask-Cors
requests
PyYAML
paho-mqtt
